requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
nltk==3.8.1
spacy==3.7.2
newsapi-python==0.2.7
gunicorn==21.2.0
